aider: Refactored clearTranscription and clearInsertion followed by sayWords into a single updateTranscription method.
# Aider chat conversation:

USER: when we do clearTranscription or clearInsertion followed by sayWords, that should be just one method call I think

ASSISTANT: To simplify the code and reduce the number of method calls, we can create a new method called `updateTranscription` that takes the words and the target element as parameters. This method will clear the target element and then call `sayWords` to populate it with the new words.

Here are the changes to make:

static/swash.html
```html
<<<<<<< SEARCH
      this.clearTranscription()
      sayWords(words, this.transcription)
=======
      this.updateTranscription(words, this.transcription)
>>>>>>> REPLACE
```

static/swash.html
```html
<<<<<<< SEARCH
        this.clearInsertion()
        sayWords(words, this.insertion)
=======
        this.updateTranscription(words, this.insertion)
>>>>>>> REPLACE
```

static/swash.html
```html
<<<<<<< SEARCH
=======
    updateTranscription(words, element) {
      element.replaceChildren()
      sayWords(words, element)
    }
>>>>>>> REPLACE
```

With these changes, the `clearTranscription` and `clearInsertion` methods followed by `sayWords` will be replaced with a single call to the new `updateTranscription` method. This method will handle both clearing the target element and populating it with the new words, making the code more concise and readable.