aider: Update other use of sayWords to use appendWordsToElement method in handleResultsEvent method of TranscriptionSession class.
# Aider chat conversation:

USER: update other use of sayWords

ASSISTANT: To update the other use of `sayWords`, we need to modify the `static/swash.html` file:

static/swash.html
```html
<<<<<<< SEARCH
          sayWords(words, span)
=======
          this.appendWordsToElement(words, span)
>>>>>>> REPLACE
```

This change replaces the call to `sayWords` with `this.appendWordsToElement` inside the `handleResultsEvent` method of the `TranscriptionSession` class.

After this modification, all uses of `sayWords` will have been refactored to use the `appendWordsToElement` method within the `TranscriptionSession` class, improving code consistency and encapsulation.